Move `default_search_path()` to oak_index
